+++ src/DotNext.Threading/Collections/Concurrent/IndexPool.cs
@@ -14,7 +14,11 @@ namespace DotNext.Collections.Concurrent;
 /// </remarks>
 public sealed class IndexPool
 {
+    private const int NoFastItem = int.MinValue;
+    private const int FrozenState = -1;
+    
     private RingBuffer<int> buffer;
+    private int fastIndex;
 
     /// <summary>
     /// Initializes a new pool of the desired capacity.
@@ -26,18 +30,19 @@ public sealed class IndexPool
         ArgumentOutOfRangeException.ThrowIfGreaterThan((uint)desiredCapacity, (uint)Array.MaxLength, nameof(desiredCapacity));
 
         buffer = new(desiredCapacity);
-        buffer.Populate();
+        fastIndex = 0;
+        buffer.Populate(startValue: 1);
     }
 
     /// <summary>
     /// Gets a value indicating that the pool is empty.
     /// </summary>
-    public bool IsEmpty => buffer.IsEmpty;
+    public bool IsEmpty => fastIndex < 0 && buffer.IsEmpty;
 
     /// <summary>
     /// Gets the capacity of the pool.
     /// </summary>
-    public int Capacity => buffer.Length;
+    public int Capacity => buffer.Length + 1;
 
     /// <summary>
     /// Tries to get the value from the pool.
@@ -46,15 +51,22 @@ public sealed class IndexPool
     /// <returns><see langword="true"/> if value is taken from the pool successfully; <see langword="false"/> is this pool is empty.</returns>
     public bool TryGet(out int value)
     {
-        ref var slot = ref buffer.TryDequeue(out var sequence);
-        if (Unsafe.IsNullRef(in slot))
+        var result = fastIndex;
+        if (result < 0 || Interlocked.CompareExchange(ref fastIndex, NoFastItem, result) != result)
         {
-            value = 0;
-            return false;
+            // slow path
+            ref var slot = ref buffer.TryDequeue(out var sequence);
+            if (Unsafe.IsNullRef(in slot))
+            {
+                value = 0;
+                return false;
+            }
+
+            result = slot.Item;
+            slot.Sequence = sequence;
         }
 
-        value = slot.Item;
-        slot.Sequence = sequence;
+        value = result;
         return true;
     }
 
@@ -68,12 +80,8 @@ public sealed class IndexPool
     /// </exception>
     public void Return(int value)
     {
-        ref var slot = ref buffer.TryEnqueue(out var sequence);
-        if (Unsafe.IsNullRef(in slot))
+        if (!TryReturn(value))
             ThrowOverflowException();
-
-        slot.Item = value;
-        slot.Sequence = sequence;
         
         [DoesNotReturn]
         [StackTraceHidden]
@@ -89,12 +97,16 @@ public sealed class IndexPool
     /// otherwise, <see langword="false"/> if this pool is frozen.</returns>
     public bool TryReturn(int value)
     {
-        ref var slot = ref buffer.TryEnqueue(out var sequence);
-        if (Unsafe.IsNullRef(in slot))
-            return false;
+        if (fastIndex >= 0 || Interlocked.CompareExchange(ref fastIndex, value, NoFastItem) is not NoFastItem)
+        {
+            ref var slot = ref buffer.TryEnqueue(out var sequence);
+            if (Unsafe.IsNullRef(in slot))
+                return false;
 
-        slot.Item = value;
-        slot.Sequence = sequence;
+            slot.Item = value;
+            slot.Sequence = sequence;
+        }
+
         return true;
     }
 
@@ -108,19 +120,23 @@ public sealed class IndexPool
     /// <see langword="true"/> if this method is called for the first time;
     /// <see langword="false"/> if the pool is already frozen.
     /// </returns>
-    public bool Freeze() => buffer.Freeze();
+    public bool Freeze()
+    {
+        fastIndex = FrozenState;
+        return buffer.Freeze();
+    }
 }
 
 file static class RingBufferExtensions
 {
-    public static void Populate(this ref RingBuffer<int> buffer)
+    public static void Populate(this ref RingBuffer<int> buffer, int startValue)
     {
         for (var index = 0; index < buffer.Length; index++)
         {
             ref var slot = ref buffer.TryEnqueue(out var sequence);
             Debug.Assert(!Unsafe.IsNullRef(in slot));
-            
-            slot.Item = index;
+
+            slot.Item = index + startValue;
             slot.Sequence = sequence;
         }
     }